Agendamento Ok, faltando tela de avaliação

+++ app/src/main/java/br/ufrpe/bsi/mpoo/petspeed/os/dominio/Triagem.java
@@ -2,15 +2,15 @@ package br.ufrpe.bsi.mpoo.petspeed.os.dominio;
 
 public class Triagem {
 
-    private double id;
+    private long id;
 
     private String sintomas;
 
-    public double getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(double id) {
+    public void setId(long id) {
         this.id = id;
     }
